refactor(MovableRangePiece): 범위 이동이 가능한 클래스 추상화
