+++ src/main/java/domain/piece/Bishop.java
@@ -1,43 +1,11 @@
 package domain.piece;
 
-import domain.directions.BishopDirections;
-import domain.directions.Direction;
 import domain.Player;
-import domain.position.Column;
-import domain.position.Position;
-import domain.position.Row;
-import java.util.ArrayList;
-import java.util.List;
+import domain.directions.BishopDirections;
 
-public class Bishop extends Piece {
-
-    private List<Position> positions = new ArrayList<>();
+public class Bishop extends MovableRangePiece {
 
     public Bishop(final Player player) {
         super(player, PieceSymbol.Bishop, new BishopDirections());
-    }
-
-    @Override
-    public List<Position> availableMovePositions(Position source) {
-        positions = new ArrayList<>();
-        for (Direction direction : directions) {
-            calculateAvailablePosition(source, direction);
-        }
-        return positions;
-    }
-
-    private void calculateAvailablePosition(final Position source, final Direction direction) {
-        int row = source.getRow() + direction.getRow();
-        int column = source.getColumn() + direction.getColumn();
-
-        while (checkOverRange(row, column)) {
-            positions.add(new Position(Row.of(row), Column.of(column)));
-            row += direction.getRow();
-            column += direction.getColumn();
-        }
-    }
-
-    private boolean checkOverRange(final int row, final int column) {
-        return Row.isRowRange(row) && Column.isColumnRange(column);
     }
 }
